Minor code refactoring

+++ Job-DB-Backup/core.ps1
@@ -26,26 +26,31 @@ function Job-DB-Backup
         $raw_db_name = ((@($DB_Backup_Prefix, $DB_Name, $DB_Backup_Suffix) | Where-Object { $_ -ne $null }) -join '_').ToString().ToLower().Replace(' ', '_')
         $raw_db_date = (Get-Date -format "ddMMyyyy_HHmmss")
 
-        Write-Output "Locating existed backups ..."
-        $raw_files_regex = "$($raw_db_name).*?\.bak"
-        $raw_files_backups = @(Get-Childitem $DB_Backup_Location | Where-Object { $_.Name -match $raw_files_regex } | Sort-Object LastWriteTime | Select -ExpandProperty Name)
-        if ($raw_files_backups.length -ge $DB_Backup_StoreCount)
+        # If parameter greater then 0 then locate existed backups and remove oldest
+        if ($DB_Backup_StoreCount -ge 1)
         {
-            $raw_files_count = $raw_files_backups.length - $DB_Backup_StoreCount
-            Write-Output "Preparing to remove $($raw_files_count + 1) backup(s) ..."
-            $raw_files_backups[0..$raw_files_count] | ForEach-Object {
-                Write-Output "Removing $($_) ..."
-                Remove-Item -Force ($DB_Backup_Location + $_)
+            Write-Output "Locate existed backups ..."
+
+            $raw_files_regex = ("{0}.*?\.bak" -f $raw_db_name)
+            $raw_files_backups = @(Get-Childitem $DB_Backup_Location | Where-Object { $_.Name -match $raw_files_regex } | Sort-Object LastWriteTime | Select -ExpandProperty Name)
+            if ($raw_files_backups.length -ge $DB_Backup_StoreCount)
+            {
+                $raw_files_count = $raw_files_backups.length - $DB_Backup_StoreCount
+                Write-Output "Prepare to remove $($raw_files_count + 1) backup(s) ..."
+                $raw_files_backups[0..$raw_files_count] | ForEach-Object {
+                    Write-Output "Remove $($_) ..."
+                    Remove-Item -Force ($DB_Backup_Location + $_)
+                }
             }
         }
 	}
 
 	process
 	{
-        $backup_name = "$($raw_db_name)_$($raw_db_date).bak"
-        $backup_query = "BACKUP DATABASE [$($DB_Name)] TO DISK = N'$($DB_Backup_Location)$($backup_name)' WITH NOFORMAT, INIT, NAME = N'$($backup_name)', SKIP, NOREWIND, NOUNLOAD, STATS = 10"
+        $backup_name = ("{0}_{1}.bak" -f $raw_db_name, $raw_db_date)
+        $backup_query = ("BACKUP DATABASE [{0}] TO DISK = N'{1}/{2}' WITH NOFORMAT, INIT, NAME = N'{2}', SKIP, NOREWIND, NOUNLOAD, STATS = 10" -f $DB_Name, $DB_Backup_Location.TrimEnd('/'), $backup_name)
 
-        Write-Output "Backuping database ..."
+        Write-Output "Backup database ..."
         Invoke-Sqlcmd -ServerInstance $DB_Server -Query $backup_query
 	}
